Expose shared IAPManager instance
